requirements.txt
Pillow==9.*
tqdm
